refactor: Rename weather example packages and aliases
